Implement hack to fix wavy underline in Chromium
`text-decoration-style` is dumb, especially when you change it after the page
is initially rendered. Firefox handles it correctly, but Chromium/Chrome does
not update the rendering box of the text and doesn't repaint on style change.
This looks really weird because sometimes the wave will be half painted, and
other times it'll get painted on hover and then never get unpainted.

The hacky workaround to this is to force the repaint of the whole box using a
hopefully invisible outline with a carefully chosen outline-offset.